Add bespoke components for enhanced editor functionality
- Introduced `BoardCanvasFrame`, `PlaceholderBox`, and `PreviewItemWrapper` as bespoke components to improve the structure and usability of the canvas boards.
- Implemented `DragNodePreview` and `FramerExpandable` for enhanced drag-and-drop functionality and smooth animations in the sidebar.
- Created `Toast` and `ToastStack` components for better toast notifications, improving user feedback during interactions.
- Updated imports across various components to utilize the new bespoke implementations, enhancing overall code organization and maintainability.

These changes significantly enhance the editor's functionality and user experience, providing a more cohesive and responsive interface.
