+++ packages/editor/app/canvas/boards/FontCollectionBoard.tsx
@@ -14,10 +14,8 @@ import { usePreview } from "@lib/hooks/use-preview"
 import { useFontCollectionBoardSpecimens } from "../hooks/use-font-collection-board-specimens"
 import { getComponentKey } from "@lib/workspace/workspace-accessors"
 import { Frame } from "@seldon/components/chrome/frames/Frame"
-import {
-  PreviewItemWrapper,
-  StyleTag,
-} from "@seldon/components/custom-components"
+import { StyleTag } from "@seldon/components/custom-components"
+import { PreviewItemWrapper } from "./PreviewItemWrapper.bespoke"
 import { CssPortal } from "../CssPortal"
 import { canvasSelectionId } from "../helpers/canvas-selection-target"
 import { BoardPreviewNode } from "./BoardPreviewNode"
