+++ packages/editor/app/canvas/boards/ThemeBoard.tsx
@@ -43,10 +43,8 @@ import {
 } from "@lib/workspace/node-tree"
 import { getComponentKey } from "@lib/workspace/workspace-accessors"
 import { Frame } from "@seldon/components/chrome/frames/Frame"
-import {
-  PreviewItemWrapper,
-  StyleTag,
-} from "@seldon/components/custom-components"
+import { StyleTag } from "@seldon/components/custom-components"
+import { PreviewItemWrapper } from "./PreviewItemWrapper.bespoke"
 import { CssPortal } from "../CssPortal"
 import { BoardPreviewNode } from "./BoardPreviewNode"
