+++ packages/editor/app/sidebars/DragNodePreview.tsx
@@ -1,3 +1,4 @@
+// BESPOKE-VIEW: hand-authored drag image for object/layer drag-and-drop.
 import { CSSProperties } from "react"
 import { ItemNode } from "@seldon/components/elements/ItemNode"
 import { IconProps } from "@seldon/components/primitives/Icon"
@@ -17,12 +18,15 @@ interface DragNodePreviewProps {
 }
 
 export function DragNodePreview({ label, icon }: DragNodePreviewProps) {
+  const icon2 = { icon }
+  const input = { value: label, readOnly: true }
+
   return (
     <ItemNode
       buttonIconic={null}
       comboboxField={{}}
-      icon2={{ icon }}
-      input={{ value: label, readOnly: true }}
+      icon2={icon2}
+      input={input}
       buttonIconic2={null}
       style={previewStyle}
     />
