+++ packages/editor/app/sidebars/objects/VMBoard.tsx
@@ -11,7 +11,7 @@ import { IndentationLevel } from "../hooks/use-indentation"
 import { useRenameInput } from "../hooks/use-rename-input"
 import { useRowBoard } from "./hooks/use-row-board"
 import { getComponentKey } from "@lib/workspace/workspace-accessors"
-import { FramerExpandable } from "@seldon/components/custom-components"
+import { FramerExpandable } from "@app/sidebars/FramerExpandable"
 import { ItemNode } from "@seldon/components/elements/ItemNode"
 import { RowSelectionTarget } from "./RowSelectionTarget"
 import { VMNode } from "./VMNode"
