+++ packages/editor/app/sidebars/objects/VMNode.tsx
@@ -14,7 +14,7 @@ import { IndentationLevel } from "../hooks/use-indentation"
 import { useRenameInput } from "../hooks/use-rename-input"
 import { useRowNode } from "./hooks/use-row-node"
 import { getNode } from "@lib/workspace/workspace-accessors"
-import { FramerExpandable } from "@seldon/components/custom-components"
+import { FramerExpandable } from "@app/sidebars/FramerExpandable"
 import { ItemNode } from "@seldon/components/elements/ItemNode"
 import { SidebarTracking } from "../../tracking/SidebarTracking"
 import { RowSelectionTarget } from "./RowSelectionTarget"
