+++ packages/editor/app/sidebars/objects/VMObjectsSidebar.tsx
@@ -16,7 +16,7 @@ import { useScrollSelection } from "./hooks/use-scroll-selection"
 import { SelectionRelationsProvider } from "./hooks/use-selection-relations"
 import { getSelectionTarget } from "@lib/workspace/selection-target"
 import { getComponentKey } from "@lib/workspace/workspace-accessors"
-import { FramerExpandable } from "@seldon/components/custom-components"
+import { FramerExpandable } from "@app/sidebars/FramerExpandable"
 import { SidebarObjects } from "@seldon/components/modules/SidebarObjects"
 import { useAddToast } from "@app/toaster/hooks/use-add-toast"
 import { BoardSection } from "../helpers/get-board-sections"
