+++ packages/editor/app/sidebars/objects/hooks/use-draggable.tsx
@@ -4,7 +4,7 @@ import { setCustomNativeDragPreview } from "@atlaskit/pragmatic-drag-and-drop/el
 import { useEffect, useRef, useState } from "react"
 import { createRoot } from "react-dom/client"
 import { Instance, Variant } from "@seldon/core"
-import { DragNodePreview } from "@seldon/components/custom-components"
+import { DragNodePreview } from "@app/sidebars/DragNodePreview"
 import { IconProps } from "@seldon/components/primitives/Icon"
 import { useNodeIcon } from "./use-node-icon"
