+++ packages/editor/app/sidebars/properties/VMProperty.tsx
@@ -3,7 +3,7 @@ import { mergeStateProps } from "@lib/views/state-props"
 import { memo } from "react"
 import { IndentationLevel } from "../hooks/use-indentation"
 import { RowPropertyProps, useRowProperty } from "./hooks/use-row-property"
-import { FramerExpandable } from "@seldon/components/custom-components"
+import { FramerExpandable } from "@app/sidebars/FramerExpandable"
 import { ComboboxFieldProps } from "@seldon/components/elements/ComboboxField"
 import { ItemProperty } from "@seldon/components/elements/ItemProperty"
 import { LayerDragRow } from "./LayerDragRow"
