+++ packages/editor/app/sidebars/properties/hooks/use-layer-draggable.tsx
@@ -4,7 +4,7 @@ import { setCustomNativeDragPreview } from "@atlaskit/pragmatic-drag-and-drop/el
 import { useEffect, useRef, useState } from "react"
 import { createRoot } from "react-dom/client"
 import { LayeredPaintKey } from "@seldon/core"
-import { DragNodePreview } from "@seldon/components/custom-components"
+import { DragNodePreview } from "@app/sidebars/DragNodePreview"
 import { IconProps } from "@seldon/components/primitives/Icon"
 
 export const LAYER_DRAG_ACTION = "properties-reorder-layer"
