+++ packages/editor/app/toaster/Toast.bespoke.tsx
@@ -1,6 +1,7 @@
 // BESPOKE-VIEW: hand-authored transitional View with inline token styling.
 // Replace with a generated workspace component once one covers the toast card.
 import { motion } from "framer-motion"
+import { CSSProperties } from "react"
 
 interface ToastProps {
   message: string
@@ -12,6 +13,18 @@ const toastVariants = {
   exit: { opacity: 0, y: 10 },
 }
 
+const toastStyle: CSSProperties = {
+  padding: "1rem 1.5rem",
+  backgroundColor: "var(--sdn-swatch-white)",
+  borderRadius: "0.75rem",
+  boxShadow:
+    "0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1)",
+  outline: "1px solid var(--sdn-swatch-black)",
+  color: "var(--sdn-swatch-black)",
+}
+
+const messageStyle: CSSProperties = { fontSize: "var(--sdn-font-size-small)" }
+
 /** Animated toast card. */
 export function Toast({ message }: ToastProps) {
   return (
@@ -21,17 +34,9 @@ export function Toast({ message }: ToastProps) {
       initial="initial"
       animate="animate"
       exit="exit"
-      style={{
-        padding: "1rem 1.5rem",
-        backgroundColor: "var(--sdn-swatch-white)",
-        borderRadius: "0.75rem",
-        boxShadow:
-          "0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1)",
-        outline: "1px solid var(--sdn-swatch-black)",
-        color: "var(--sdn-swatch-black)",
-      }}
+      style={toastStyle}
     >
-      <p style={{ fontSize: "var(--sdn-font-size-small)" }}>{message}</p>
+      <p style={messageStyle}>{message}</p>
     </motion.div>
   )
 }
