+++ packages/editor/app/toaster/Toaster.tsx
@@ -3,7 +3,8 @@
 import { AnimatePresence } from "framer-motion"
 import { useShallow } from "zustand/react/shallow"
 import { useToastStore } from "./hooks/use-toast-store"
-import { Toast, ToastStack } from "@seldon/components/custom-components"
+import { Toast } from "./Toast.bespoke"
+import { ToastStack } from "./ToastStack.bespoke"
 
 export function Toasts() {
   const toasts = useToastStore(useShallow((state) => state.toasts))
