+++ packages/editor/seldon/custom-components/index.ts
@@ -13,17 +13,9 @@ export { StyleTag } from "./primitives/StyleTag"
 export { Box } from "./containers/Box"
 export { Backdrop } from "./containers/Backdrop"
 export { PositionedPanel } from "./containers/PositionedPanel"
-export { FloatingPanelSurface } from "./containers/FloatingPanelSurface"
 export { DropzoneSurface } from "./containers/DropzoneSurface"
-export { BoardCanvasFrame } from "./containers/BoardCanvasFrame"
-export { PreviewItemWrapper } from "./containers/PreviewItemWrapper"
-export { PlaceholderBox } from "./containers/PlaceholderBox"
 export { CanvasOutline } from "./indicators/CanvasOutline"
 export { InsertIndicatorLine } from "./indicators/InsertIndicatorLine"
 export { Pointer } from "./indicators/Pointer"
 export { OverlayLayer } from "./indicators/OverlayLayer"
 export { PlacementZoneSurface } from "./indicators/PlacementZoneSurface"
-export { ToastStack } from "./containers/ToastStack"
-export { Toast } from "./feedback/Toast"
-export { FramerExpandable } from "./containers/FramerExpandable"
-export { DragNodePreview } from "./containers/DragNodePreview"
